requirements.txt
black==23.1.0
click==8.1.3
python-dotenv==1.0.0
isort==5.12.0
openai==0.27.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
python-telegram-bot[rate-limiter]==20.1
ruff==0.0.259
mypy-extensions
tomli==2.0.1
typing_extensions==4.5.0
